python: Deduplicate telescoping logic
